+++ client/Packages/com.beamable/Common/Runtime/Api/Auth/AuthApi.cs
@@ -9,7 +9,6 @@ using UnityEngine;
 
 namespace Beamable.Common.Api.Auth
 {
-
 	public enum CredentialUsageStatus
 	{
 		// the credential is "available" to be used.
@@ -17,6 +16,10 @@ namespace Beamable.Common.Api.Auth
 
 		// this means that the credential is already bound to a playerId
 		ASSIGNED_TO_AN_ACCOUNT,
+		/// <summary>
+		/// This could mean one of two things - either the credentials provided were incorrect or the identity provider is not configured properly.
+		/// The former can happen while integrating third party identity providers.
+		/// </summary>
 		INVALID_CREDENTIAL
 	}
 
@@ -62,7 +65,7 @@ namespace Beamable.Common.Api.Auth
 			{
 				var resp = await _requester
 					.Request<AvailabilityResponse>(Method.GET, $"{ACCOUNT_URL}/available?email={encodedEmail}", null,
-					                               false);
+						false);
 
 				if (resp.available)
 				{
@@ -89,10 +92,10 @@ namespace Beamable.Common.Api.Auth
 			try
 			{
 				var resp = await _requester
-				       .Request<AvailabilityResponse>(
-					       Method.GET,
-					       $"{ACCOUNT_URL}/available/third-party{qb}", null,
-					       false);
+					.Request<AvailabilityResponse>(
+						Method.GET,
+						$"{ACCOUNT_URL}/available/third-party{qb}", null,
+						false);
 
 				if (resp.available)
 				{
@@ -946,8 +949,20 @@ namespace Beamable.Common.Api.Auth
 	[Serializable]
 	public class ExternalAuthenticationResponse
 	{
+	   /// <summary>
+	   /// The user ID in the external system (wallet ID, OAuth ID, etc.)
+	   /// When provided, it indicates that the federation service identified the user and provided his external ID.
+	   /// </summary>
 		public string user_id;
+		/// <summary>
+		/// The challenge associated with the external authentication.
+		/// When provided, it indicates that the external authentication is pending verification.
+		/// </summary>
 		public string challenge;
+		/// <summary>
+		/// The time-to-live (TTL) of the challenge.
+		/// When provided, it indicates that the external authentication is pending verification.
+		/// </summary>
 		public int challenge_ttl;
 	}